requirements.txt
django==3.2
pillow==8.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0